Improved determination of the maximum index. (#1)
This avoids relying on a lowest placeholder; any matrix value that is equal to
the placeholder would not trigger the update code, meaning that the maximum
index would be incorrect. We also just report zero for the maximum value in the
case of an empty matrix, given that the result should not be used anyway.

Also added some tests to check for correct behavior in the empty matrix.