Streamlined the BlockThresholds definition by re-using Thresholds.
This allows us to easily create a BlockThresholds from an existing Thresholds,
which cuts down on redundant code. It also supports applications that need to
return a constant-typed object from code that may or may not use blocks; in
such cases, we can just promote a Thresholds object to BlockThresholds.