Accept optional workspaces/buffers in find_median_mad overloads.
This allows us to eliminate an explicit overload, as missing workspaces/buffers
are indicated by a NULL pointer that is allocated inside the function. We also
add an overload for the single-block case with a const input pointer, which
avoids callers having to remember to explicitly copy it into a new buffer.